% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraWrapper.R
\name{ProduceGenesetJSONfile}
\alias{ProduceGenesetJSONfile}
\title{ProduceGenesetJSONfile}
\usage{
ProduceGenesetJSONfile(spectraDefaultGeneSets, immportGeneSetsPath, outputpath)
}
\arguments{
\item{spectraDefaultGeneSets}{The path to the default spectra library.}

\item{immportGeneSetsPath}{The path to the RDS file containing the ImmPort gene sets.}

\item{outputpath}{The path to the output JSON file, which will be provided to Spectra.}
}
\description{
A helper function to prepare a gene-set JSON file compatible with Spectra.
}
