% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraWrapper.R
\name{RunSpectra}
\alias{RunSpectra}
\title{RunSpectra}
\usage{
RunSpectra(
  seuratObj,
  seuratToAnnDataDir,
  assayName,
  inputJSONfile,
  immportfile
)
}
\arguments{
\item{seuratObj}{Input Seurat object}

\item{seuratToAnnDataDir}{Directory to house files used in generating Ann Data}

\item{assayName}{Assay used for generating Ann Data}

\item{inputJSONfile}{Input Seurat object}

\item{immportfile}{Input Seurat object}
}
\description{
Wrapper function for the Spectra package
}
