% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraWrapper.R
\name{SeuratToAData}
\alias{SeuratToAData}
\title{SeuratToAData}
\usage{
SeuratToAData(seuratObj, seuratToAdataDir, assayName = "RNA")
}
\arguments{
\item{seuratObj}{The Seurat object to be written into a 10x file format}

\item{seuratToAdataDir}{The local path to write the output files.}

\item{assayName}{The name of the gene expression data assay.}
}
\description{
A wrapper function to prepare a Seurat object for Conga.
}
