% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraWrapper.R
\name{SpectraToSeuratMetadata}
\alias{SpectraToSeuratMetadata}
\title{SpectraToSeuratMetadata}
\usage{
SpectraToSeuratMetadata(seuratObj, cellresults_file, plotComponents = TRUE)
}
\arguments{
\item{seuratObj}{The Seurat object to which to add the Spectra results.}

\item{cellresults_file}{The path to the Spectra cell scores file.}

\item{plotComponents}{Boolean controlling whether to plot the Spectra components.}
}
\description{
A helper function to add Spectra results to a Seurat object as metadata.
}
