% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraWrapper.R
\name{SpectraToSeuratReduction}
\alias{SpectraToSeuratReduction}
\title{SpectraToSeuratReduction}
\usage{
SpectraToSeuratReduction(
  seuratObj,
  seuratToAdataDir,
  assayName = "RNA",
  reduction.name = "spectra",
  reduction.key = "Spectra_"
)
}
\arguments{
\item{seuratObj}{The Seurat object to which to add the Spectra results.}

\item{seuratToAdataDir}{Intermediate directory containing the Spectra results.}

\item{assayName}{The name of the gene expression data assay.}

\item{reduction.name}{The name of the reduction to add to the Seurat object.}

\item{reduction.key}{The key to use for the reduction.}
}
\description{
A helper function to add Spectra results to a Seurat object as a reduction.
}
