% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraWrapper.R
\name{SpectraWrapper}
\alias{SpectraWrapper}
\title{SpectraWrapper}
\usage{
SpectraWrapper(seuratObj)
}
\arguments{
\item{seuratObj}{Input Seurat object}
}
\description{
Wrapper function for the Spectra package
}
